api/requirements.txt
Flask==2.0.1

# chineseocr_lite
onnxruntime==1.7.0
opencv-python==4.5.2.52
Pillow==8.2.0
pyclipper==1.2.1
Shapely==1.7.1

# easyocr
easyocr==1.3.1

# paddleocr
imgaug==0.4.0
lmdb==1.2.1
paddlepaddle==2.1.0
